+++ backend/resources/views/admin/product/editproductcategory.blade.php
@@ -114,7 +114,7 @@
                 },
                 category_order: {
                     validators: {
-                        integer: {
+                        numeric: {
                         	message: 'The value is not an integer'
                     	}
                     }
